requirements.txt
APScheduler==3.6.3
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
codecov==2.1.12
colorama==0.4.4
coverage==5.5
csscompressor==0.9.5
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2.7
django-appconf==1.0.5
django-compressor==2.4.1
django-crispy-forms==1.13.0
django-rest-framework==0.1.0
django-webpack-loader==1.4.0
djangorestframework==3.12.4
filelock==3.2.0

gunicorn==20.1.0
httplib2==0.20.0
idna==3.2
joblib==1.0.1
keras==2.6.0
mock==4.0.3
nltk==3.6.3
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pbr==5.6.0
Pillow==8.3.2
platformdirs==2.4.0
pluggy==1.0.0
protobuf==3.18.0
psycopg2-binary==2.9.2
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-fsutil==0.5.0
python-telegram-bot==13.7
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
rcssmin==1.0.6
regex==2021.9.24
requests==2.26.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
tensorflow @ https://storage.googleapis.com/tensorflow/mac/cpu/tensorflow-1.0.0-py3-none-any.whl
toml==0.10.2
tornado==6.1
tox==3.24.4
tqdm==4.62.3
tzdata==2021.2.post0
tzlocal==4.0
uritemplate==3.0.1
urllib3==1.26.7
virtualenv==20.8.1
whitenoise==5.3.0
